feat: created layout of new UI, fixed width, created search-category pannel
