+++ src/common/config/appearance-options.ts
@@ -39,8 +39,8 @@ export const defaultAppearanceOptions: AppearanceOptions = {
     showSearchIcon: true,
     showSearchResultNumbers: false,
     smoothScrolling: true,
-    userInputHeight: 60,
-    windowWidth: 600,
+    userInputHeight: 200,
+    windowWidth: 1024,
     userInputBorderRadius: "0px",
     userInputFontSize: "24px",
     userInputFontWeight: "200",
