+++ src/renderer/renderer.ts
@@ -3,6 +3,7 @@ import { VueEventChannels } from "./vue-event-channels";
 import { vueEventDispatcher } from "./vue-event-dispatcher";
 import { userInputComponent } from "./user-input-component";
 import { searchResultsComponent } from "./search-results-component";
+import { searchCategoriesPannel } from "./search-categories-pannel";
 import { ipcRenderer } from "electron";
 import { IpcChannels } from "../common/ipc-channels";
 import { SearchResultItem } from "../common/search-result-item";
@@ -65,6 +66,7 @@ import { loremIpsumSettingsComponent } from "./settings/lorem-ipsum-settings-com
 
 Vue.component("user-input", userInputComponent);
 Vue.component("search-results", searchResultsComponent);
+Vue.component("search-categories", searchCategoriesPannel);
 Vue.component("settings", settingsComponent);
 Vue.component("general-settings", generalSettingsComponent);
 Vue.component("appearance-settings", appearanceSettingsComponent);
