+++ styles/app.css
@@ -145,7 +145,30 @@ body {
     overflow: hidden;
 }
 
-.search-results__item {
+.search-results__container {
+    width: 100%;
+    height: 100%;
+    display: grid;
+    grid-template-columns: repeat(12, 1fr);
+    column-gap: var(--search-results--item-info-container-padding);
+}
+.search-results-categories {
+    grid-column: 1 / span 3;
+    height: 100%;
+    overflow-y: auto;
+    overflow-x: hidden;
+    position: relative;
+}
+.search-results-categories.scroll-disabled {
+    overflow: hidden;
+}
+.search-results-pannel {
+    grid-column: 4 / span 5;
+}
+.search-preview-pannel {
+    grid-column: 9 / span 4;
+}
+.search-categories__item {
     display: grid;
     grid-template-columns: auto 1fr auto;
     align-items: center;
@@ -157,22 +180,22 @@ body {
     user-select: none;
 }
 
-.search-results__item.active {
+.search-categories__item.active {
     background-color: var(--search-results--item-active-background-color);
 }
 
-.search-results__item-icon-container {
+.search-categories__item-icon-container {
     width: var(--search-results--item-icon-size);
     height: var(--search-results--item-icon-size);
     position: relative;
     flex-shrink: 0;
     opacity: 0.75;
 }
-.search-results__item-icon-container.active {
+.search-categories__item-icon-container.active {
     opacity: 1;
 }
 
-.search-results__item-icon {
+.search-categories__item-icon {
     width: 100%;
     height: 100%;
     display: flex;
@@ -181,34 +204,34 @@ body {
     align-items: center;
 }
 
-.search-results__item-icon--url {
+.search-categories__item-icon--url {
     max-width: 100%;
     max-height: 100%;
 }
 
-.search-results__item-icon--svg {
+.search-categories__item-icon--svg {
     width: 100%;
     height: 100%;
     fill: var(--search-results--item-name-text-color);
 }
 
-.search-results__item-icon--svg.active {
+.search-categories__item-icon--svg.active {
     fill: var(--search-results--item-active-text-color);
 }
 
-.search-results__item-icon--svg > svg {
+.search-categories__item-icon--svg > svg {
     width: var(--search-results--item-icon-size);
     height: var(--search-results--item-icon-size);
 }
 
-.search-results__item-icon--color {
+.search-categories__item-icon--color {
     height: var(--search-results--item-icon-size);
     width: var(--search-results--item-icon-size);
     border-radius: 50%;
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.25);
 }
 
-.search-results__item-info-container {
+.search-categories__item-info-container {
     padding: 0 var(--search-results--item-info-container-padding);
     box-sizing: border-box;
     display: flex;
@@ -217,47 +240,47 @@ body {
     overflow: hidden;
 }
 
-.search-results__item-number-container {
+.search-categories__item-number-container {
     color: var(--search-results--item-name-text-color);
     font-size: var(--search-results--item-search-result-number-font-size);
     opacity: 0.5;
 }
 
-.search-results__item-number-container.active {
+.search-categories__item-number-container.active {
     color: var(--search-results--item-active-text-color);
 }
 
-.search-results__item-name,
-.search-results__item-description {
+.search-categories__item-name,
+.search-categories__item-description {
     width: 100%;
     overflow: hidden;
     text-overflow: ellipsis;
     white-space: nowrap;
 }
 
-.search-results__item-name {
+.search-categories__item-name {
     color: var(--search-results--item-name-text-color);
     font-size: var(--search-results--item-name-font-size);
     font-weight: var(--search-results--item-name-font-weight);
     padding: var(--search-results--item-name-padding);
 }
 
-.search-results__item-name.active {
+.search-categories__item-name.active {
     color: var(--search-results--item-active-text-color);
 }
 
-.search-results__item-description {
+.search-categories__item-description {
     display: none;
     color: var(--search-results--item-description-text-color);
     font-size: var(--search-results--item-description-font-size);
     font-weight: var(--search-results--item-description-font-weight);
 }
 
-.search-results__item-description.active {
+.search-categories__item-description.active {
     color: var(--search-results--item-active-description-color);
 }
 
-.search-results__item-description.visible {
+.search-categories__item-description.visible {
     display: block;
 }
